// tools/generators/_ui-component/schema.d.ts
export interface Schema {
  path: string;
  name: string;
  prefix: string;
}

export interface ExtendedSchema extends Schema {
  name: string;
  className: string;
  propertyName: string;
  constantName: string;
  fileName: string;
  nameInSingular: string;
  classNameInSingular: string;
  propertyNameInSingular: string;
  constantNameInSingular: string;
  fileNameInSingular: string;
}